Try to resolve dependancy loop
